noaa_updates/requirements_noaa.txt
# Dependências para módulo NOAA
aiohttp>=3.8.0
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.28.0
psutil>=5.9.0
scipy>=1.9.0
customtkinter>=5.2.0
Pillow>=9.0.0
python-dateutil>=2.8.2
